Add per-version integration classes
Each NodeInfo version now has its own integration class that:
- Registers its version to the endpoint enum
- Adds its discovery and JRD links
- Provides version-specific schema
- Implements version-specific data via filters

This allows other plugins to add new NodeInfo versions or modify
existing ones by hooking into the appropriate filters.

New filters:
- nodeinfo_versions: Register supported versions
- nodeinfo_discovery_links: Add discovery document links
- nodeinfo_jrd_links: Add WebFinger/Host-Meta links
- nodeinfo_schema: Modify the JSON schema
